pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
server = "server.app:main"

[project]
name = "email-triage-env"
version = "0.4.0"
description = "Customer Support Email Triage OpenEnv Environment"
dependencies = [
    "pydantic>=2.7.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "openenv-core>=0.2.0",
    "openai>=1.20.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
